Toronto Fitness Club Frontend/PF-group_10905/PB/requirements.txt
django==4.1.3
Pillow==9.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-extensions==3.2.1
django-filter==22.1
django-cors-headers==3.13.0
geocoder
haversine
python-dateutil